Seeder に Factory クラスを使用する

+++ database/migrations/2021_10_20_0921752_create_okrs_table.php
@@ -16,7 +16,7 @@ class CreateOkrsTable extends Migration
         Schema::create('okrs', function (Blueprint $table) {
             $table->increments('id');
             $table->text('name')->comment('目標');
-            $table->integer('score')
+            $table->float('score')
                 ->nullable()
                 ->comment('総合スコア');
             $table->integer('user_id')
